+++ src/test/scala/com/gu/HelloSpec.scala
@@ -1,0 +1,9 @@
+package com.gu
+
+import org.scalatest._
+
+class MainSpec extends FlatSpec with Matchers {
+  "The Main object" should "satisfy criteria" in {
+    true shouldEqual true
+  }
+}
