--- src/test/scala/example/HelloSpec.scala
@@ -1,9 +1,0 @@
-package example
-
-import org.scalatest._
-
-class HelloSpec extends FlatSpec with Matchers {
-  "The Hello object" should "say hello" in {
-    Hello.greeting shouldEqual "hello"
-  }
-}
